Fix annoying runtime in station trait (#89425)
## About The Pull Request

For some reason I was getting this station trait a lot while testing and
it would runtime on my usual setup where I spawn myself in from
observer, because there were no bank accounts in existence.
This won't have any particular impact on a real round because you are
probably not getting a real round with no bank accounts.

## Changelog

Not player facing