[NO GBP] Remove the call to `install_node.sh` from the Windows CI build script (#89494)
## About The Pull Request
Back when I moved all of the CI to use `actions/setup-node` in
https://github.com/tgstation/tgstation/pull/89363, it seems I forgot to
remove a line from the Windows build script that attempts to call
`tools/ci/install_node.sh`.

Thankfully, nothing was broken - it just complains that the script
doesn't exist, and continues on as if nothing happened.

That said, there's also no point to the error - so this PR removes that
line.